requirements.txt
streamlit==1.39.0
pandas==2.2.3
requests==2.32.3
plotly==5.24.1
python-dateutil==2.9.0.post0
bcrypt==4.0.1